restructured "roots" module
